MAINT: interpolate: move RGI tests to a separate file
